컴포넌트명 수정, 브랜드 컴포넌트 추가
기존 "브랜드" -> "홈 스타일링"으로 컴포넌트명 수정
라우터 수정 및 브랜드 컴포넌트 추가
