+++ src/assets/images/brand/index.ts
@@ -16,6 +16,11 @@ import henredn03 from "./henredn03.png"
 import umasqu01 from "./umasqu01.png"
 import umasqu02 from "./umasqu02.png"
 import umasqu03 from "./umasqu03.png"
+import alfdn01 from "./alfdn01.png"
+import alfdn02 from "./alfdn02.png"
+import alfdn03 from "./alfdn03.png"
+import alfdn04 from "./alfdn04.png"
+import alfdn05 from "./alfdn05.png"
 
 export {
   jamesLee,
@@ -35,4 +40,9 @@ export {
   umasqu01,
   umasqu02,
   umasqu03,
+  alfdn01,
+  alfdn02,
+  alfdn03,
+  alfdn04,
+  alfdn05,
 }
