+++ src/pages/style/Brand.tsx
@@ -1,116 +1,77 @@
 import React, { useState } from "react"
 import { Box, FormControl, Typography, Checkbox, Button } from "@mui/material"
-import WhiteTitle from '../../components/title/WhiteTitle'
-import { onerain03 } from "../../assets/images/product"
-import { IconRadioOn, IconRadioOff } from "../../assets/images"
-import onerainDetail from "../../assets/images/brand/onerainDetail.png"
-import { Link } from "react-router-dom";
+import { alfdn01, alfdn02, alfdn03, alfdn04, alfdn05 } from "../../assets/images/brand"
 import NavigationBar from "../../components/NavigationBar"
-import { chair02, table01 } from "../../assets/images/product"
+import List from "../like/Like/List"
+import LeftTitle from "../../components/title/LeftTitle"
 
-export const brandItemData = [
+export const testData = [
   {
     id: 0,
-    imgsrc: onerain03,
-    shopName: 'ONERAIN',
-    modelName: '[23 Series] OR-512',
-    productName: '원레인 더티 러그',
-    price: 230000
+    imgsrc: alfdn01,
+    like: true,
+    likeCount: 556,
+    shopName: 'ALFDN',
+    modelName: '[SET] AD-012',
+    productName: '알프든 테이블 도자기 세트',
+    price: 230000,
+    colors: ['#CBC6C3']
   },
   {
     id: 1,
-    imgsrc: chair02,
-    shopName: 'ONERAIN',
-    modelName: '[23 Series] OR-309',
-    productName: '원레인 아트체어',
-    price: 230000
+    imgsrc: alfdn02,
+    like: false,
+    likeCount: 556,
+    shopName: 'ALFDN',
+    modelName: '[23Series] AD-316',
+    productName: '알프든 체어',
+    price: 1230000,
+    colors: ['#282629', '#CBC6C3']
   },
   {
     id: 2,
-    imgsrc: table01,
-    shopName: 'ONERAIN',
-    modelName: '[23 Series] OR-555',
-    productName: '원레인 테이블',
-    price: 420000
+    imgsrc: alfdn03,
+    like: false,
+    likeCount: 556,
+    shopName: 'ALFDN',
+    modelName: '[SET] AD-322',
+    productName: 'Wooden Table, Chair set',
+    price: 830000,
+    colors: ['#3F2916', '#D67531']
+  },
+  {
+    id: 3,
+    imgsrc: alfdn04,
+    like: false,
+    likeCount: 556,
+    shopName: 'ALFDN',
+    modelName: '[23Series] AD-405',
+    productName: 'Basic sofa',
+    price: 2430000,
+    colors: ['#192552', '#A6A8A7']
   },
 ]
 
 const Brand = () => {
-  function formatPrice(price: any) {
-    return price.toLocaleString();
-  }
-  const [selectedIconIndex, setSelectedIconIndex] = useState(0);
-  const iconCoordinates = [
-    { top: "500px", left: "200px" }, // 아이콘 0의 좌표
-    { top: "430px", left: "90px" },  // 아이콘 1의 좌표
-    { top: "340px", left: "210px" }, // 아이콘 2의 좌표
-  ];
-
-  const handleIconClick = (index: any) => {
-    setSelectedIconIndex(index);
-  };
-
-  const selectedIconCoordinates = iconCoordinates[selectedIconIndex] || {}; // 선택된 아이콘의 좌표
-
-  // 선택된 아이콘의 데이터 가져오기
-  const selectedBrandData = brandItemData[selectedIconIndex] || {};
-
+  
   return (
-    <>
-    <Box sx={{position:"relative", height: "100vh"}}>
-      <Box sx={{position:"absolute", top:0, zIndex:2,width: "100%"}}>
-        <WhiteTitle title={["모던한 인테리어 디자인"]}/>
+    <Box sx={{minHeight:"100vh", position:"relative", display:"flex", flexDirection:"column", justifyContent:"space-between"}}>
+      <Box>
+        <LeftTitle title={["브랜드"]} />
+        <Box sx={{ position: "relative", overflow: 'hidden' }}>
+          <Box sx={{ position: "absolute", width: "100%", display: "flex", flexDirection: "column", alignItems: "center", top: "55px" }}>
+            <Typography sx={{ fontSize: 20, fontWeight: "700", color: "#FFFFFF" }}>ALFDN</Typography>
+          </Box>
+          <Box sx={{ width: "100%", display: "flex" }}>
+            <img src={alfdn05} alt="" style={{ objectFit: "cover", width: "100%" }} />
+          </Box>
+        </Box>
+        <List data={testData}/>
       </Box>
-      <img src={onerainDetail} alt="" style={{position:"absolute", width: "100%", zIndex:1, objectFit: "cover", height: "100%"}}/>
-      {iconCoordinates.map((coord, index) => (
-        <div
-          key={index}
-          style={{
-            zIndex: 2,
-            position: "absolute",
-            top: coord.top,
-            left: coord.left,
-            cursor: "pointer",
-            opacity: selectedIconIndex === index ? 0 : 1, // 선택된 아이콘에 대한 스타일 적용
-          }}
-          onClick={() => handleIconClick(index)}
-        >
-          <IconRadioOff />
-        </div>
-      ))}
-      {selectedIconCoordinates && (
-        <div
-          style={{
-            zIndex: 2,
-            position: "absolute",
-            top: selectedIconCoordinates.top,
-            left: selectedIconCoordinates.left,
-          }}
-        >
-          <IconRadioOn />
-        </div>
-      )}
+      <Box sx={{position:"sticky", bottom:0, width:"100%"}}>
+        <NavigationBar />
+      </Box>
     </Box>
-    
-      {selectedBrandData && (
-        <Link to={"/shop/product_view"}>
-        <Box sx={{position:"fixed", bottom: "74px", backgroundColor: "rgba(255, 255, 255, 0.8)", width: "100%", p: 2, display: "flex", zIndex:2}}>
-        <Box sx={{pr: 2, borderRight: "1px solid #666666", display: "flex"}}>
-          <img src={selectedBrandData.imgsrc} alt="" style={{width: "100px", height: "100px", objectFit: "cover", zIndex:10, opacity: 1, }}/>
-        </Box>
-        <Box sx={{display:"flex", flexDirection:"column", justifyContent:"center", pl:2}}>
-          <Typography sx={{ fontSize: 14, fontWeight: "500", lineHeight:"20px", letterSpacing: "2%", color:"#333333" }}>{selectedBrandData.shopName}</Typography>
-          <Typography sx={{ fontSize: 12, fontWeight: "400", lineHeight:"20px", letterSpacing: "-0.25px", color:"#999999" }}>{selectedBrandData.modelName}</Typography>
-          <Typography sx={{ fontSize: 12, fontWeight: "400", lineHeight:"20px", letterSpacing: "-0.25px", color:"#999999" }}>{selectedBrandData.productName}</Typography>
-          <Typography sx={{ fontSize: 12, fontWeight: "400", lineHeight: "20px", letterSpacing: "-0.25px", color: "#666666" }}>
-            {selectedBrandData.price.toLocaleString()} won
-          </Typography>
-        </Box>
-      </Box>
-      </Link>
-      )}
-    <NavigationBar/>
-  </>
   )
 }
